Finally got it to work
Added dependencies

Added new directory for package

changed src to be src root
