--- Cap_Proj/src/Student.java
@@ -1,101 +1,0 @@
-import java.util.ArrayList;
-import java.util.List;
-
-public class Student {
-    String name;
-    String studentID;
-    String major;
-    String semesterAdmittedToProgram;
-    List<Course> courses;
-
-    public Student(String name, String studentID, String major, String semesterAdmittedToProgram) {
-        this.name = name;
-        this.studentID = studentID;
-        this.major = major;
-        this.semesterAdmittedToProgram = semesterAdmittedToProgram;
-        this.courses = new ArrayList<>();
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getStudentID() {
-        return studentID;
-    }
-
-    public void setStudentID(String studentID) {
-        this.studentID = studentID;
-    }
-
-    public String getMajor() {
-        return major;
-    }
-
-    public void setMajor(String major) {
-        this.major = major;
-    }
-
-    public String getSemesterAdmittedToProgram() {
-        return semesterAdmittedToProgram;
-    }
-
-    public void setSemesterAdmittedToProgram(String semesterAdmittedToProgram) {
-        this.semesterAdmittedToProgram = semesterAdmittedToProgram;
-    }
-
-    public List<Course> getCourses() {
-        return courses;
-    }
-
-    public void setCourses(List<Course> courses) {
-        this.courses = courses;
-    }
-
-    public void addCourse(Course course) {
-        courses.add(course);
-    }
-    public void applyAdditionalRules() {
-        removeCsIppAssignment();
-        handleElectiveCourses();
-    }
-
-    private void removeCsIppAssignment() {
-        courses.removeIf(course -> course.getCourseCode().equals("5177") && course.getCourseName().equals("CS IPP ASSIGNMENT"));
-    }
-
-    private void handleElectiveCourses() {
-        Course electiveCourse = null;
-        List<Course> coursesToRemove = new ArrayList<>();
-
-        for (Course course : courses) {
-            if (course.getCourseCode().equals("5333") || course.getCourseCode().equals("5343") || course.getCourseCode().equals("5348")) {
-                if (electiveCourse == null) {
-                    electiveCourse = course;
-                } else if (course.getGradeValue() > electiveCourse.getGradeValue()) {
-                    coursesToRemove.add(electiveCourse);
-                    electiveCourse = course;
-                } else if (course.getGradeValue() == electiveCourse.getGradeValue()) {
-                    if (course.getCourseCode().equals("5343")) {
-                        coursesToRemove.add(electiveCourse);
-                        electiveCourse = course;
-                    } else if (course.getCourseCode().equals("5348") && !electiveCourse.getCourseCode().equals("5343")) {
-                        coursesToRemove.add(electiveCourse);
-                        electiveCourse = course;
-                    } else {
-                        coursesToRemove.add(course);
-                    }
-                } else {
-                    coursesToRemove.add(course);
-                }
-            }
-        }
-
-        courses.removeAll(coursesToRemove);
-    }
-    // Add any other necessary methods
-}
