+++ Cap_Proj/src/cap_proj/PdfReader.java
@@ -1,3 +1,5 @@
+package cap_proj;
+
 import java.io.File;
 import java.io.IOException;
 import org.apache.pdfbox.pdmodel.PDDocument;
